Refactor health handling: move HandleHealth function to a new file and update usage in enclave service
